+++ test/shared.js
@@ -1,33 +1,45 @@
 import crypto from 'crypto'
 import DB from '../src/lib/Db'
-import config from '../src/lib/config'
+import { makeConfig, config as userConfig } from '../src/lib/config'
+import defaultConfig from '../src/lib/defaultConfig'
+import collections from '../src/lib/collections'
 import { getDbBlocksCollections } from '../src/lib/blocksCollections'
 import NativeContracts from '../src/lib/NativeContracts'
 import initConfig from '../src/lib/initialConfiguration'
 import { addrTypes } from '../src/lib/types'
+import net from 'net'
 
+export const config = makeConfig()
 export const nativeContracts = NativeContracts(initConfig)
 const testDatabase = 'dbToTest'
 
+const getDbName = config => config.db.database
+
+const writeOnlyDbs = [getDbName(userConfig), getDbName(defaultConfig)]
+
 export const testDb = ({ dbName } = {}) => {
   dbName = dbName || testDatabase
+  if (writeOnlyDbs.includes(dbName)) throw new Error(`Don't use production databases to test!!!`)
   const dbConf = Object.assign(config.db, { database: dbName })
   const database = new DB(dbConf)
+  database.setLogger(null)
   let db
   const getDb = async () => {
     if (!db) db = await database.db()
     return db
   }
   const dropDb = () => getDb().then(db => db.dropDatabase())
-  return Object.freeze({ getDb, dropDb })
+  return Object.freeze({ getDb, dropDb, db: database, config })
 }
 
-export const testCollections = async (dropDb) => {
-  const database = testDb()
+export const testCollections = async (dropDb, database) => {
+  database = database || testDb()
   if (dropDb) await database.dropDb()
   const db = await database.getDb()
-  const collections = await getDbBlocksCollections(db)
-  return collections
+  const names = defaultConfig.collectionsNames
+  await database.db.createCollections(collections, { names })
+  const colls = await getDbBlocksCollections(db)
+  return colls
 }
 
 export const fakeBlocks = (count = 10, { max, addTimestamp } = {}) => {
@@ -44,21 +56,26 @@ export const fakeBlocks = (count = 10, { max, addTimestamp } = {}) => {
 
 export const randomBlockHash = () => '0x' + crypto.randomBytes(32).toString('hex')
 
-export const randomBlockNumber = (max) => {
-  max = max || 10 ** 6
-  return Math.floor(Math.random() * max)
-}
+export const randomNumber = (max, min = 1) => Math.floor(Math.random() * max) + min
 
-export const fakeBlock = (max) => {
-  let number = randomBlockNumber(max)
-  let hash = randomBlockHash()
-  return { hash, number }
+export const randomBlockNumber = (max, min = 0) => {
+  max = max || 10 ** 6
+  return randomNumber(max, min)
 }
 
 export const fakeTx = (transactionIndex, { hash, number }) => {
   let blockHash = hash || randomBlockHash()
   let blockNumber = number || randomBlockNumber()
   return { blockHash, blockNumber, transactionIndex }
+}
+
+export const fakeBlock = (max) => {
+  let number = randomBlockNumber(max)
+  let hash = randomBlockHash()
+  let miner = randomAddress()
+  let txs = randomNumber(10)
+  let transactions = [...Array(txs)].map(() => randomBlockHash())
+  return { hash, number, miner, transactions }
 }
 
 export const randomAddress = () => `0x${crypto.randomBytes(20).toString('hex')}`
@@ -89,4 +106,26 @@ export function Spy (obj, method) {
     obj[method] = org
   }
   return Object.freeze({ spy, remove })
+}
+
+export function isPortInUse (port, host) {
+  host = host || '127.0.0.1'
+  if (isNaN(port)) throw new Error('Port must be a number')
+  return new Promise((resolve, reject) => {
+    const server = net.createServer()
+      .once('error', err => (err.code === 'EADDRINUSE' ? resolve(true) : reject(err)))
+      .once('listening', () => {
+        server.once('close', () => resolve(false))
+        server.close()
+      })
+      .listen(port, host)
+  })
+}
+
+export function asyncWait (time = 1000) {
+  return new Promise((resolve) => setTimeout(resolve, time))
+}
+
+export function isRejected (promise) {
+  return promise.then(() => false).catch(err => Promise.resolve(() => { throw err }))
 }
